refactor: keep lifecycle contract projection explicit
